+++ tests/tests_pytorch/deprecated_api/test_remove_2-0.py
@@ -300,62 +300,3 @@ def test_v2_0_0_deprecated_mc_save_checkpoint():
         match=r"ModelCheckpoint.save_checkpoint\(\)` was deprecated in v1.6 and is no longer supported as of 1.8.",
     ):
         mc.save_checkpoint(trainer)
-
-
-@pytest.mark.parametrize(
-    "attribute",
-    [
-        "gpus",
-        "num_gpus",
-        "root_gpu",
-        "devices",
-        "tpu_cores",
-        "ipus",
-        "use_amp",
-        "weights_save_path",
-        "lightning_optimizers",
-        "should_rank_save_checkpoint",
-        "validated_ckpt_path",
-        "tested_ckpt_path",
-        "predicted_ckpt_path",
-        "verbose_evaluate",
-    ],
-)
-def test_v2_0_0_trainer_unsupported_getters(attribute):
-    trainer = Trainer()
-    with pytest.raises(
-        AttributeError, match=f"`Trainer.{attribute}` was deprecated in v1.6 and is no longer accessible as of v1.8."
-    ):
-        getattr(trainer, attribute)
-
-
-@pytest.mark.parametrize(
-    "attribute",
-    [
-        "validated_ckpt_path",
-        "tested_ckpt_path",
-        "predicted_ckpt_path",
-        "verbose_evaluate",
-    ],
-)
-def test_v2_0_0_trainer_unsupported_setters(attribute):
-    trainer = Trainer()
-    with pytest.raises(
-        AttributeError, match=f"`Trainer.{attribute}` was deprecated in v1.6 and is no longer accessible as of v1.8."
-    ):
-        setattr(trainer, attribute, None)
-
-
-def test_v2_0_0_lightning_module_unsupported_use_amp():
-    model = BoringModel()
-    with pytest.raises(
-        AttributeError,
-        match="`LightningModule.use_amp` was deprecated in v1.6 and is no longer accessible as of v1.8.",
-    ):
-        _ = model.use_amp
-
-    with pytest.raises(
-        AttributeError,
-        match="`LightningModule.use_amp` was deprecated in v1.6 and is no longer accessible as of v1.8.",
-    ):
-        model.use_amp = False
