+++ tests/tests_pytorch/graveyard/test_core.py
@@ -1,0 +1,31 @@
+# Copyright The PyTorch Lightning team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import pytest
+
+from pytorch_lightning.demos.boring_classes import BoringModel
+
+
+def test_v2_0_0_lightning_module_unsupported_use_amp():
+    model = BoringModel()
+    with pytest.raises(
+        AttributeError,
+        match="`LightningModule.use_amp` was deprecated in v1.6 and is no longer accessible as of v1.8.",
+    ):
+        _ = model.use_amp
+
+    with pytest.raises(
+        AttributeError,
+        match="`LightningModule.use_amp` was deprecated in v1.6 and is no longer accessible as of v1.8.",
+    ):
+        model.use_amp = False
